refactor all pre-game activities, fragments and adapters into /game/prep
