+++ app/src/androidTest/java/com/github/fribourgsdp/radio/GameSettingsActivityTest.kt
@@ -16,6 +16,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.data.Playlist
+import com.github.fribourgsdp.radio.game.prep.*
 import com.github.fribourgsdp.radio.mockimplementations.MockGameSettingsActivity
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
