+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockGameSettingsActivity.kt
@@ -2,11 +2,11 @@ package com.github.fribourgsdp.radio.mockimplementations
 
 import android.content.Context
 import android.os.Bundle
-import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.data.Genre
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.data.Song
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.game.prep.GameSettingsActivity
 import org.mockito.Mockito.*
 
 class MockGameSettingsActivity: GameSettingsActivity() {
