+++ app/src/androidTest/java/com/github/fribourgsdp/radio/mockimplementations/MockJoinWithQRCodeFragment.kt
@@ -2,18 +2,9 @@ package com.github.fribourgsdp.radio.mockimplementations
 
 import android.app.Activity
 import android.content.Context
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
-import android.widget.Toast
-import androidx.core.os.bundleOf
-import androidx.fragment.app.setFragmentResult
-import com.github.fribourgsdp.radio.JoinWithQRCodeFragment
+import com.github.fribourgsdp.radio.game.prep.JoinWithQRCodeFragment
 import com.github.fribourgsdp.radio.R
 import eu.livotov.labs.android.camview.ScannerLiveView
-import io.agora.rtc.IRtcEngineEventHandler
 import org.mockito.Mockito
 
 class MockJoinWithQRCodeFragment(ctx: Context, activity: Activity) : JoinWithQRCodeFragment(ctx, activity) {
