+++ app/src/main/java/com/github/fribourgsdp/radio/MainActivity.kt
@@ -13,6 +13,8 @@ import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.config.SettingsActivity
 import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.data.view.UserProfileActivity
+import com.github.fribourgsdp.radio.game.prep.GameSettingsActivity
+import com.github.fribourgsdp.radio.game.prep.JoinGameActivity
 
 class MainActivity : MyAppCompatActivity() {
     private val db = FirestoreDatabase()
